--- a/docs/site/public/schemas/components/status-light.json
+++ b/docs/site/public/schemas/components/status-light.json
@@ -27,7 +27,12 @@
         "magenta",
         "fuchsia",
         "purple",
-        "seafoam"
+        "seafoam",
+        "pink",
+        "turquoise",
+        "cinnamon",
+        "brown",
+        "silver"
       ],
       "default": "informative"
     },