--- a/tools/diff-generator/src/lib/store-output.js
+++ b/tools/diff-generator/src/lib/store-output.js
@@ -17,12 +17,52 @@ export default function storeOutput(filePath, output) {
   try {
     const outputDirectory = filePath.slice(0, filePath.lastIndexOf(path.sep));
     if (!existsSync(outputDirectory)) {
-      mkdirSync(outputDirectory, { recursive: true });
+      try {
+        mkdirSync(outputDirectory, { recursive: true });
+      } catch (dirError) {
+        if (dirError.code === "EACCES") {
+          throw new Error(
+            `Permission denied creating directory "${outputDirectory}". Check directory permissions.`,
+          );
+        } else if (dirError.code === "ENOTDIR") {
+          throw new Error(
+            `Cannot create directory "${outputDirectory}": a file with this name already exists.`,
+          );
+        } else {
+          throw new Error(
+            `Failed to create output directory "${outputDirectory}": ${dirError.message}`,
+          );
+        }
+      }
     }
 
-    writeFileSync(filePath, output);
+    try {
+      writeFileSync(filePath, output);
+    } catch (writeError) {
+      if (writeError.code === "EACCES") {
+        throw new Error(
+          `Permission denied writing to file "${filePath}". Check file and directory permissions.`,
+        );
+      } else if (writeError.code === "ENOSPC") {
+        throw new Error(`Not enough disk space to write file "${filePath}".`);
+      } else if (writeError.code === "EISDIR") {
+        throw new Error(
+          `Cannot write to "${filePath}": it is a directory, not a file.`,
+        );
+      } else {
+        throw new Error(
+          `Failed to write output file "${filePath}": ${writeError.message}`,
+        );
+      }
+    }
   } catch (error) {
-    console.log(`FAILED TO WRITE OUTPUT FILE: ${filePath}`);
-    console.error(error);
+    const contextualError = error.message.includes("Failed to")
+      ? error.message
+      : `Output file operation failed for "${filePath}": ${error.message}`;
+
+    console.error(contextualError);
+
+    // Re-throw to allow calling code to handle appropriately
+    throw error;
   }
 }