… .changeset/performance-optimization.md
… .cursor/mcp.json
… docs/site/public/schemas/components/meter.json
… docs/site/public/schemas/components/popover.json
… docs/site/public/schemas/components/search-field.json
… docs/site/public/schemas/components/select-box.json
… docs/site/public/schemas/components/text-area.json
… docs/site/public/schemas/components/text-field.json
… docs/site/public/schemas/token-types/token.json
… docs/site/public/schemas/token-types/typography.json
… packages/component-schemas/schemas/components/meter.json
… pnpm-lock.yaml
… tools/diff-generator/OPTIMIZATION_INTEGRATION.md
… tools/diff-generator/PERFORMANCE_ANALYSIS.md
… tools/diff-generator/README.md
… tools/diff-generator/optimization-plan.md
… tools/diff-generator/package.json
… tools/diff-generator/test/fileImportRefactored.test.js
… tools/diff-generator/test/storeOutput.test.js
… tools/diff-generator/test/temp-merge-tokens/packages/tokens/src/colors1.json
… tools/diff-generator/test/temp-merge-tokens/packages/tokens/src/colors2.json
… tools/diff-generator/test/temp-merge-tokens/pnpm-lock.yaml
… tools/diff-generator/test/temp-tokens/packages/tokens/src/color-1.json
… tools/diff-generator/test/temp-tokens/packages/tokens/src/color-2.json
… tools/diff-generator/test/temp-tokens/pnpm-lock.yaml
… tools/optimized-diff/README.md
… tools/optimized-diff/moon.yml
… tools/optimized-diff/package.json
… tools/optimized-diff/src/benchmark.js
… tools/optimized-diff/src/engine.js