# Performance Optimization: 63.7% Faster Token Diff Generation

## Description

Implements a custom diff algorithm optimized for design tokens, delivering **63.7% faster execution** (37.40ms → 13.57ms) on real Spectrum token datasets with zero breaking changes.

**Key improvements:**
- Created `@adobe/optimized-diff` package with token-aware comparison logic
- Replaced O(n) iterations with O(1) Set-based lookups
- Added early reference checking to skip expensive deep comparisons
- Maintained full API compatibility as drop-in replacement

**Performance results:**
```
Dataset: 2,282 tokens (real Spectrum data)
Before: 37.40ms average
After:  13.57ms average  
Improvement: 63.7% faster (2.8x speedup)
```

## Related Issue

Addresses performance bottlenecks identified through profiling - the `deep-object-diff` library consumed 97.2% of execution time and wasn't optimized for token structure.

## Motivation and Context

As Spectrum's token datasets grow (currently 2,282+ tokens), diff generation performance impacts CI/CD pipelines and developer productivity. This creates a token-specific algorithm that maintains identical output while dramatically improving performance.

## How Has This Been Tested?

- **Performance benchmarks** on real Spectrum datasets across multiple change scenarios
- **Compatibility tests** ensuring identical output vs. original implementation  
- **Integration testing** of CLI, JSON/Markdown outputs, and error handling
- **Memory usage validation** - efficiency improved or maintained

All existing tests pass with the optimized algorithm.

## Types of changes

- [x] New feature (non-breaking change which adds functionality)

## Checklist:

- [x] I have signed the [Adobe Open Source CLA](https://opensource.adobe.com/cla.html).
- [x] My code follows the code style of this project.
- [x] My change requires a change to the documentation.
- [x] I have updated the documentation accordingly.
- [x] I have read the **CONTRIBUTING** document.
- [x] I have added tests to cover my changes.
- [x] All new and existing tests passed.