@GarthDB 2025-08-01
<!-- moon-run-report: unknown -->

## Run report for [51df27a3](https://github.com/adobe/spectrum-tokens/commit/51df27a35654f9dfaaf8bf4ca912bc287d611703)
Total time: 27.8s | Comparison time: 30.7s | Estimated savings: 3s (9.7% faster)
|     | Action | Time | Status | Info |
| :-: | :----- | ---: | :----- | :--- |
| ⬛️ | `SetupProto(0.51.4)` | 0.1ms | Skipped |  |
| 🟩 | `SyncWorkspace` | 7.7ms | Passed |  |
| 🟩 | `SyncProject(site)` | 0.4ms | Passed |  |
| 🟩 | `SyncProject(diff)` | 0.5ms | Passed |  |
| 🟩 | `RunTask(site:copyTokenTypeSchema)` | 654ms | Passed |  |
| 🟩 | `RunTask(site:copyComponentSchema)` | 665.6ms | Passed |  |
| ⬛️ | `SetupToolchain(node:~20.12)` | 1.4s | Skipped |  |
| 🟩 | `InstallWorkspaceDeps(node:~20.12)` | 6.7s | Passed |  |
| 🟩 | `RunTask(diff:diff)` | 179.7ms | Passed |  |
| 🟩 | `RunTask(site:clean)` | 218.7ms | Passed |  |
| 🟩 | `RunTask(diff:test)` | 5.3s | Passed |  |
| 🟩 | `RunTask(site:build)` | 19.6s | Passed |  |
| 🟩 | `RunTask(site:export)` | 76.1ms | Passed |  |

<details><summary><strong>Touched files</strong></summary><div>

```
.cursor/mcp.json
docs/site/public/schemas/components/meter.json
docs/site/public/schemas/components/popover.json
docs/site/public/schemas/components/search-field.json
docs/site/public/schemas/components/select-box.json
docs/site/public/schemas/components/status-light.json
docs/site/public/schemas/components/text-area.json
docs/site/public/schemas/components/text-field.json
docs/site/public/schemas/token-types/token.json
docs/site/public/schemas/token-types/typography.json
pnpm-lock.yaml
tools/diff-generator/OPTIMIZATION_INTEGRATION.md
tools/diff-generator/PERFORMANCE_ANALYSIS.md
tools/diff-generator/README.md
tools/diff-generator/optimization-plan.md
tools/diff-generator/package.json
tools/diff-generator/src/lib/cli.js
tools/diff-generator/src/lib/diff.js
tools/diff-generator/src/lib/file-import.js
tools/diff-generator/src/lib/formatterHandlebars.js
tools/diff-generator/src/lib/optimized-diff.js
tools/diff-generator/src/lib/store-output.js
tools/diff-generator/test/fileImportRefactored.test.js
tools/diff-generator/test/formatterHandlebars.test.js
tools/diff-generator/test/storeOutput.test.js
tools/optimized-diff/README.md
tools/optimized-diff/ava.config.js
tools/optimized-diff/moon.yml
tools/optimized-diff/package.json
tools/optimized-diff/src/benchmark.js
tools/optimized-diff/src/engine.js
tools/optimized-diff/src/helpers.js
tools/optimized-diff/src/index.js
tools/optimized-diff/test/engine.test.js
```

</div></details>